Validating TF models

+++ pkg/types/spec/api.go
@@ -50,9 +50,9 @@ type LocalModelCache struct {
 
 type CuratedModelResource struct {
 	userconfig.ModelResource
-	S3Path    bool  `json:"s3_path"`
-	ZipFormat bool  `json:"zip_format"`
-	Versions  []int `json:"versions"`
+	S3Path    bool    `json:"s3_path"`
+	ZipFormat bool    `json:"zip_format"`
+	Versions  []int64 `json:"versions"`
 }
 
 func GetAPISpec(apiConfig *userconfig.API, models []CuratedModelResource, projectID string, deploymentID string) *API {
